Test case debugging and near match
The code was fully modified to have functionality of the point source
scatter problem. The test case has been verified to match the analytical
solution using the field point solutions. The only discrepancy that can
be seen is the value of phi on the boundary. The output is twice that of
the analytical solution, so this error should still be resolved.